fix qf,ac
