+++ service/src/main/java/com/guru/service/parser/impl/ParserQF.java
@@ -8,10 +8,12 @@ import com.guru.domain.model.Trip;
 import com.guru.domain.repository.QueryRepository;
 import com.guru.parser.impl.qfparser.QFParser;
 import com.guru.service.parser.interf.ParserActor;
+import com.guru.vo.temp.Account;
 import com.guru.vo.transfer.RequestData;
 import com.guru.vo.transfer.Status;
 import com.guru.vo.transfer.StatusCount;
 import com.guru.vo.utils.ProcessRequestHelperService;
+import org.apache.http.impl.client.DefaultHttpClient;
 
 import javax.inject.Inject;
 import java.util.Collection;
@@ -33,7 +35,6 @@ public class ParserQF extends UntypedActor implements ParserActor {
     @Inject
     private QueryRepository queryRepository;
 
-    //для всех
 
     LoggingAdapter log = Logging.getLogger(getContext().system(), this);
 
@@ -48,6 +49,11 @@ public class ParserQF extends UntypedActor implements ParserActor {
             RequestData requestData = (RequestData) message;
             RequestData newRequestData = new RequestData(requestData);
 
+            Account account = com.guru.vo.temp.AccountUtils.getAccount("AC");
+            DefaultHttpClient httpclient = QFParser.login("1924112640", "Kin", "4152");
+
+            newRequestData.setAccount(account);
+            newRequestData.setDefaultHttpClient(httpclient);
             if (Objects.equals(newRequestData.getType(), "ow")) {
                 List<Date> owDates = newRequestData.getOwDates();
                 for (Date date : owDates) {
@@ -63,10 +69,9 @@ public class ParserQF extends UntypedActor implements ParserActor {
                         float status = (float) statusCount.getCurrentStatus() / statusCount.getMaxStatus() * 100;
                         queryRepository.updateStatus(queryId, (int) status);
                         System.out.println(status + "%");
-                        System.out.println(Status.count++);
+                        System.out.println("Count -" + Status.count++);
                         //System.out.println("Count - " + count++);
                     }
-                    ;
 
                 }
             } else {
@@ -74,50 +79,21 @@ public class ParserQF extends UntypedActor implements ParserActor {
                         .getDaysBetweenDates(newRequestData.getOw_start_date(), newRequestData.getOw_end_date());
                 List<Date> rtDates = ProcessRequestHelperService
                         .getDaysBetweenDates(newRequestData.getRt_start_date(), newRequestData.getRt_end_date());
-                for (Date date : owDates) {
-                    newRequestData.setOw_end_date(date);
-                    newRequestData.setOw_start_date(date);
+                for (int i = 0; i < owDates.size(); i++) {
+                    newRequestData.setOw_start_date(owDates.get(i));
+                    newRequestData.setRt_start_date(rtDates.get(i));
                     Collection<Trip> flights = qfParser.parse(newRequestData);
                     if (flights.size() != 0) {
                         processingResultOfParserActor.tell(flights, self());
-                       // System.out.println(Status.count++);
-                    }
-                    else {
+                    } else {
                         Long queryId = (long) newRequestData.getRequest_id();
                         StatusCount statusCount = Status.getStatusCountByQueryId(queryId);
                         Status.updateStatus(queryId);
                         float status = (float) statusCount.getCurrentStatus() / statusCount.getMaxStatus() * 100;
                         queryRepository.updateStatus(queryId, (int) status);
                         System.out.println(status + "%");
-                        System.out.println(Status.count++);
-                        //System.out.println("Count - " + count++);
+                        System.out.println("Count -" + Status.count++);
                     }
-
-                }
-                String destination = newRequestData.getDestination();
-                String origin = newRequestData.getOrigin();
-                newRequestData.setOrigin(destination);
-                newRequestData.setDestination(origin);
-                for (Date date : rtDates) {
-                    newRequestData.setOw_start_date(date);
-                    newRequestData.setOw_end_date(date);
-                    Collection<Trip> flights = qfParser.parse(newRequestData);
-                    if (flights.size() != 0) {
-                        processingResultOfParserActor.tell(flights, self());
-
-                    }
-                    else {
-                        Long queryId = (long) newRequestData.getRequest_id();
-                        StatusCount statusCount = Status.getStatusCountByQueryId(queryId);
-                        Status.updateStatus(queryId);
-                        float status = (float) statusCount.getCurrentStatus() / statusCount.getMaxStatus() * 100;
-                        queryRepository.updateStatus(queryId, (int) status);
-                        System.out.println(status + "%");
-                        System.out.println("Count -"+Status.count++);
-                        //System.out.println("Count - " + count++);
-                    }
-
-
                 }
             }
         } else unhandled(message);
